Code refactoring & interaction
It wasn't the best idea not to make consecutive commits, but it is what it is, and now there are a lot of changes in one commit.
The interaction system has been redesigned for the zone trigger

+++ Assets/Scripts/Interfaces/IAttackable.cs
@@ -1,5 +1,5 @@
 ﻿public interface IAttackable
 {
-     void GetHit(float damage,float pushForce);
+     void GetHit(int damage,float pushForce);
 }
